nonebot_plugin_setu_customization/pyproject.toml
[tool.poetry]
name = "nonebot_plugin_setu_customization"
version = "1.3.1"
description = "Nonebot2 可动态管理API并带网页浏览的setu插件"
authors = ["nikissXI <1299577815@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_setu_customization"}]
homepage = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_setu_customization"
repository = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_setu_customization"
documentation = "https://github.com/nikissXI/nonebot_plugins/tree/main/nonebot_plugin_setu_customization#README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0-rc.1"
nonebot-adapter-onebot = "^2.0.0"
httpx = "^0.23.0"
httpx-socks = "^0.7.3"
beautifulsoup4 = "^4.0.0"
Pillow = "^9.0.0"
ujson = "^5.0.0"
starlette = "^0.19.0"
apscheduler = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
